requirements.txt
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
cramjam==2.11.0
dotenv==0.9.9
entsoe-py==0.7.8
fastparquet==2024.11.0
fsspec==2025.10.0
idna==3.11
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pyarrow==21.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
